♻️ Refactor model.py into data_model directory
