+++ backend/data_model/career.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+class Career(BaseModel):
+    role: str
+    company: str
+    period: str
+    description: str
