+++ backend/data_model/club.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+class Club(BaseModel):
+    name: str
+    period: str
+    description: str
