+++ backend/data_model/education.py
@@ -1,0 +1,7 @@
+from typing import Optional
+from pydantic import BaseModel
+
+class Education(BaseModel):
+    name: str
+    period: str
+    description: str
